+++ src/meltano/__init__.py
@@ -1,2 +1,2 @@
 # Managed by bumpversion
-__version__ = "1.14.0"
+__version__ = "1.14.3"