+++ src/meltano/api/controllers/errors.py
@@ -1,0 +1,5 @@
+class InvalidFileNameError(Exception):
+    """Occurs when an invalid file name is provided."""
+
+    def __init__(self, name):
+        self.name = name