+++ src/meltano/core/plugin/settings_service.py
@@ -21,11 +21,11 @@ class PluginSettingMissingError(Error):
         super().__init__(f"Cannot find setting {name} in {plugin}")
 
 
-class PluginSettingValueSource(int, Enum):
-    ENV = 0
-    MELTANO_YML = 1
-    DB = 2
-    DEFAULT = 3
+class PluginSettingValueSource(str, Enum):
+    ENV = "env"  # 0
+    MELTANO_YML = "meltano_yml"  # 1
+    DB = "db"  # 2
+    DEFAULT = "default"  # 3
 
 
 class PluginSettingsService:
@@ -49,13 +49,42 @@ class PluginSettingsService:
 
         return {k: v for k, v in values.items() if v != REDACTED_VALUE}
 
-    def as_config(
+    def profile_with_config(
+        self, session, plugin: PluginRef, profile: Profile, redacted=False
+    ):
+        plugin_install = self.get_install(plugin)
+
+        plugin_install.use_profile(profile)
+
+        full_config = self.config_with_sources(
+            session, plugin_install, redacted=redacted
+        )
+
+        return {
+            **profile.canonical(),
+            "config": {key: config["value"] for key, config in full_config.items()},
+            "config_sources": {
+                key: config["source"] for key, config in full_config.items()
+            },
+        }
+
+    def profiles_with_config(
+        self, session, plugin: PluginRef, redacted=False
+    ) -> List[Dict]:
+        plugin_install = self.get_install(plugin)
+
+        return [
+            self.profile_with_config(session, plugin, profile, redacted=redacted)
+            for profile in (Profile.DEFAULT, *plugin_install.profiles)
+        ]
+
+    def config_with_sources(
         self,
         session,
         plugin: PluginRef,
         sources: List[PluginSettingValueSource] = None,
         redacted=False,
-    ) -> Dict:
+    ):
         # defaults to the meltano.yml for extraneous settings
         plugin_install = self.get_install(plugin)
         plugin_def = self.get_definition(plugin)
@@ -72,29 +101,14 @@ class PluginSettingsService:
             if redacted and value and setting.kind == "password":
                 value = REDACTED_VALUE
 
-            config[setting.name] = value
+            config[setting.name] = {"value": value, "source": source}
 
         return config
 
-    def as_profile_configs(
-        self, session, plugin: PluginRef, redacted=False
-    ) -> List[Dict]:
-        plugin_install = self.get_install(plugin)
+    def as_config(self, *args, **kwargs) -> Dict:
+        full_config = self.config_with_sources(*args, **kwargs)
 
-        profiles = []
-        for profile in (Profile.DEFAULT, *plugin_install.profiles):
-            plugin_install.use_profile(profile)
-
-            profiles.append(
-                {
-                    **profile.canonical(),
-                    "config": self.as_config(
-                        session, plugin_install, redacted=redacted
-                    ),
-                }
-            )
-
-        return profiles
+        return {key: config["value"] for key, config in full_config.items()}
 
     def as_env(
         self, session, plugin: PluginRef, sources: List[PluginSettingValueSource] = None