+++ src/webapp/src/utils/constants.js
@@ -1,0 +1,2 @@
+export const ENV = 'env'
+export const MELTANO_YML = 'meltano_yml'